@@ -2,15 +2,15 @@
 	"name": "kmfrontend",
 	"private": true,
 	"dependencies": {
-		"@cospired/i18n-iso-languages": "^4.1.0",
-		"@sentry/react": "^7.93.0",
-		"antd": "5.13.2",
+		"@cospired/i18n-iso-languages": "^4.2.0",
+		"@sentry/react": "^7.102.1",
+		"antd": "5.14.2",
 		"dayjs": "^1.11.10",
 		"diff2html": "^3.4.47",
-		"dompurify": "^3.0.8",
+		"dompurify": "^3.0.9",
 		"filesize": "^10.1.0",
-		"i18n-iso-countries": "^7.7.0",
-		"i18next": "^23.7.16",
+		"i18n-iso-countries": "^7.10.0",
+		"i18next": "^23.10.0",
 		"lodash": "^4.17.21",
 		"minireset.css": "^0.0.7",
 		"path-browserify": "^1.0.1",
@@ -20,14 +20,14 @@
 		"react-audio-player": "^0.17.0",
 		"react-beautiful-dnd": "^13.1.1",
 		"react-dom": "18.2.0",
-		"react-i18next": "^14.0.0",
+		"react-i18next": "^14.0.5",
 		"react-image-crop": "^11.0.5",
-		"react-router-dom": "6.21.2",
+		"react-router-dom": "6.22.1",
 		"react-text-transition": "^3.1.0",
-		"react-toastify": "^10.0.3",
-		"react-virtuoso": "^4.6.2",
+		"react-toastify": "^10.0.4",
+		"react-virtuoso": "^4.7.1",
 		"resize-observer-polyfill": "^1.5.1",
-		"sass": "^1.69.7",
+		"sass": "^1.71.1",
 		"slug": "^8.2.3",
 		"socket.io-client": "4.7.4",
 		"use-async-memo": "^1.2.5",
@@ -36,25 +36,25 @@
 	"devDependencies": {
 		"@fortawesome/fontawesome-free": "^6.5.1",
 		"@types/lodash": "^4.14.202",
-		"@types/react": "18.2.48",
+		"@types/react": "18.2.58",
 		"@types/react-aria-menubutton": "^6.2.14",
 		"@types/react-beautiful-dnd": "^13.1.8",
-		"@types/react-dom": "18.2.18",
-		"@typescript-eslint/eslint-plugin": "^6.19.0",
-		"@typescript-eslint/parser": "^6.19.0",
+		"@types/react-dom": "18.2.19",
+		"@typescript-eslint/eslint-plugin": "^7.0.2",
+		"@typescript-eslint/parser": "^7.0.2",
 		"@vitejs/plugin-react": "^4.2.1",
-		"eslint": "^8.56.0",
+		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-react": "^7.33.2",
-		"eslint-plugin-security": "^2.1.0",
-		"eslint-plugin-simple-import-sort": "^10.0.0",
+		"eslint-plugin-security": "^2.1.1",
+		"eslint-plugin-simple-import-sort": "^12.0.0",
 		"eslint-plugin-you-dont-need-lodash-underscore": "^6.13.0",
 		"i18next-scanner": "^4.4.0",
-		"prettier": "^3.2.4",
+		"prettier": "^3.2.5",
 		"source-map-explorer": "^2.5.3",
 		"typescript": "5.3.3",
-		"vite": "^5.0.11",
-		"vite-plugin-node-polyfills": "^0.19.0"
+		"vite": "^5.1.4",
+		"vite-plugin-node-polyfills": "^0.21.0"
 	},
 	"scripts": {
 		"analyze": "source-map-explorer 'build/static/js/*.js'",
