@@ -172,46 +174,47 @@
     "@types/i18next-fs-backend": "^1.1.5",
     "@types/ip": "^1.1.3",
     "@types/js-yaml": "^4.0.9",
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/langs": "^2.0.4",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/langs": "^2.0.5",
     "@types/lodash": "4.14.202",
     "@types/mocha": "^10.0.6",
     "@types/multer": "^1.4.11",
-    "@types/node": "20.11.6",
-    "@types/pg": "^8.11.0",
+    "@types/node": "20.11.24",
+    "@types/pg": "^8.11.2",
     "@types/pg-copy-streams": "^1.2.5",
     "@types/qrcode": "^1.5.5",
-    "@types/randomstring": "^1.1.11",
+    "@types/randomstring": "^1.1.12",
     "@types/readline-sync": "^1.4.8",
-    "@types/semver": "^7.5.6",
+    "@types/semver": "^7.5.8",
     "@types/slug": "^5.0.7",
     "@types/tmi.js": "^1.8.6",
     "@types/uuid": "^9.0.8",
     "@types/which": "^3.0.3",
     "@types/yesql": "^4.1.4",
-    "@typescript-eslint/eslint-plugin": "^6.19.1",
-    "@typescript-eslint/parser": "^6.19.1",
-    "chai": "^5.0.3",
-    "electron": "^28.2.0",
-    "electron-builder": "^24.9.1",
-    "esbuild": "^0.19.12",
-    "eslint": "^8.56.0",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "chai": "^5.1.0",
+    "electron": "^29.1.0",
+    "electron-builder": "^24.12.0",
+    "esbuild": "^0.20.1",
+    "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-security": "^2.1.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "husky": "^9.0.5",
+    "eslint-plugin-security": "^2.1.1",
+    "eslint-plugin-simple-import-sort": "^12.0.0",
+    "husky": "^9.0.11",
     "i18next-scanner": "^4.4.0",
-    "lint-staged": "^15.2.0",
-    "mocha": "^10.2.0",
+    "lint-staged": "^15.2.2",
+    "mocha": "^10.3.0",
     "postgrator-cli": "^7.0.0",
-    "prettier": "3.2.4",
+    "prettier": "3.2.5",
     "rimraf": "^5.0.5",
     "ts-node": "^10.9.2",
-    "typescript": "5.3.3"
+    "typescript": "5.3.3",
+    "vitest": "^1.3.1"
   },
   "optionalDependencies": {
     "x11": "^2.3.0"
