@@ -23,6 +23,7 @@ import CheckedKaraMenuModal from '../modals/CheckedKaraMenuModal';
 import PlaylistCommandsModal from '../modals/PlaylistCommandsModal';
 import ActionsButtons from './ActionsButtons';
 import KaraReviews from './KaraReviews';
+import { getPlaylistIcon } from '../../../utils/playlist';
 
 interface IProps {
 	side: 'left' | 'right';
@@ -73,37 +74,12 @@ function PlaylistHeader(props: IProps) {
 		props.onChangeTags(tagType, value);
 	};
 
-	const getPlaylistIcon = (playlist: PlaylistElem) => {
-		// public & current playlist :  play-circle & globe icons
-		if (playlist?.flag_public && playlist?.flag_current) return ['fa-play-circle', 'fa-globe'];
-		// public playlist : globe icon
-		if (playlist?.flag_public) return ['fa-globe'];
-		// current playlist : play-circle icon
-		if (playlist?.flag_current) return ['fa-play-circle'];
-		// library : book icon
-		if (playlist?.plaid === nonStandardPlaylists.library) return ['fa-book'];
-		// animelist depending of user settings
-		if (playlist?.plaid === nonStandardPlaylists.animelist)
-			return [`icon-${context?.globalState.settings.data.user.anime_list_to_fetch}`];
-		// blacklist : ban icon
-		if (playlist?.plaid === context.globalState.settings.data.state.blacklistPlaid) return ['fa-ban'];
-		// whitelist : check-circle icon
-		if (playlist?.plaid === context.globalState.settings.data.state.whitelistPlaid) return ['fa-check-circle'];
-		// fallback playlist : arrows-turn-to-dots icon
-		if (playlist?.plaid === context.globalState.settings.data.state.fallbackPlaid)
-			return ['fa-arrows-turn-to-dots'];
-		// favorites : star icon
-		if (playlist?.plaid === nonStandardPlaylists.favorites) return ['fa-star'];
-		// others playlist : list-ol icon
-		return ['fa-list-ol'];
-	};
-
 	const getListToSelect = () => {
 		return props.playlistList.map(playlist => {
 			return {
 				value: playlist?.plaid,
 				label: playlist?.name,
-				icons: getPlaylistIcon(playlist),
+				icons: getPlaylistIcon(playlist, context),
 			};
 		});
 	};
