@@ -216,11 +216,15 @@ export async function isUserAllowedToAddKara(plaid: string, user: User, duration
 	try {
 		const conf = getConfig();
 		if (+conf.Karaoke.Quota.Type === 0) return true;
+		const playlistsToConsider =
+			conf?.Karaoke?.Quota?.FreeAcceptedSongs === true
+				? [plaid]
+				: [getState().publicPlaid, getState().currentPlaid, plaid];
 		let limit: number;
 		switch (+conf.Karaoke.Quota.Type) {
 			case 2:
 				limit = conf.Karaoke.Quota.Time;
-				let time = await selectSongTimeSpentForUser(plaid, user.login);
+				let time = await selectSongTimeSpentForUser(playlistsToConsider, user.login);
 				if (!time) time = 0;
 				if (limit - time - duration < 0) {
 					logger.debug(
@@ -235,7 +239,7 @@ export async function isUserAllowedToAddKara(plaid: string, user: User, duration
 			case 1:
 			default:
 				limit = conf.Karaoke.Quota.Songs;
-				const count = await selectSongCountForUser(plaid, user.login);
+				const count = await selectSongCountForUser(playlistsToConsider, user.login);
 				if (count >= limit) {
 					logger.debug(`User ${user.login} tried to add more songs than he/she was allowed (${limit})`, {
 						service,
@@ -721,17 +725,17 @@ export async function addKaraToPlaylist(params: AddKaraParams) {
 		const duplicateCheckList =
 			user.type === 0
 				? // Admin can add a song multiple times in the current or any other playlist, even by the same user
-				  conf.Playlist.AllowDuplicates
+					conf.Playlist.AllowDuplicates
 					? []
 					: // Option to allow is not set : removing duplicates from songs to add
-					  plContents
+						plContents
 				: // Not an admin adding these songs.
-				  conf.Playlist.AllowPublicDuplicates === 'allowed'
-				  ? // A user isn't allowed to add the same song twice
-				    // But several users can add the same song twice
-				    // So we remove all songs they haven't added already from the list that's going to be compared later.
-				    plContents.filter(plc => plc.username === requester)
-				  : plContents;
+					conf.Playlist.AllowPublicDuplicates === 'allowed'
+					? // A user isn't allowed to add the same song twice
+						// But several users can add the same song twice
+						// So we remove all songs they haven't added already from the list that's going to be compared later.
+						plContents.filter(plc => plc.username === requester)
+					: plContents;
 		karaList = karaList.filter(k => !duplicateCheckList.map(plc => plc.kid).includes(k.kid));
 
 		profile('addKaraToPL-checkDuplicates');
@@ -882,7 +886,11 @@ export async function copyKaraToPlaylist(plc_ids: number[], plaid: string, pos?:
 		}
 		for (const plc of plcs) {
 			// If source is public playlist and destination current playlist, free up PLCs from the public playlist.
-			if (plc.plaid === getState().publicPlaid && plaid === getState().currentPlaid) {
+			if (
+				plc.plaid === getState().publicPlaid &&
+				plaid === getState().currentPlaid &&
+				getConfig().Karaoke.Quota.FreeAcceptedSongs
+			) {
 				PLCsToFree.push(plc.plcid);
 			}
 			plc.added_at = new Date();
