@@ -1417,26 +1417,6 @@ export async function pushCommits(repoName: string, push: Push, ignoreFTP?: bool
 				if (media.old === null || media.old === media.new) {
 					const path = await resolveFileInDirs(media.new, resolvedPathRepos('Medias', repoName));
 					await ftp.upload(path[0]);
-				} else if (media.new === null) {
-					// Deleted file
-					try {
-						await ftp.delete(media.old);
-					} catch (err) {
-						logger.warn(`File ${media.old} could not be deleted on FTP`, { service });
-					}
-				} else if (media.new !== media.old) {
-					// Renamed file or new upload with different sizes, let's find out!
-					if (media.sizeDifference) {
-						const path = await resolveFileInDirs(media.new, resolvedPathRepos('Medias', repoName));
-						await ftp.upload(path[0]);
-						try {
-							await ftp.delete(media.old);
-						} catch (err) {
-							logger.warn(`File ${media.old} could not be deleted on FTP`, { service });
-						}
-					} else {
-						await ftp.rename(basename(media.old), basename(media.new));
-					}
 				}
 			}
 			await ftp.disconnect();
@@ -1461,12 +1441,45 @@ export async function pushCommits(repoName: string, push: Push, ignoreFTP?: bool
 						await git.rm(removedFile);
 					}
 				}
-				await git.commit(commit.message);
+				if (!push.squash) await git.commit(commit.message);
 				task.incr();
+			}
+			if (push.squash) {
+				await git.commit(push.squash);
 			}
 			// All our commits are hopefully done. Just in case we'll update repository now.
 			await updateGitRepo(repoName);
 			await git.push();
+			// Let's do deletes and renames on FTP now. And pray it doesn't fail.
+			if (!ignoreFTP && push.modifiedMedias.length > 0) {
+				await ftp.connect();
+				for (const media of push.modifiedMedias) {
+					if (media.old === null || media.old === media.new) {
+						// Upload, do nothing
+					} else if (media.new === null) {
+						// Deleted file
+						try {
+							await ftp.delete(media.old);
+						} catch (err) {
+							logger.warn(`File ${media.old} could not be deleted on FTP`, { service });
+						}
+					} else if (media.new !== media.old) {
+						// Renamed file or new upload with different sizes, let's find out!
+						if (media.sizeDifference) {
+							const path = await resolveFileInDirs(media.new, resolvedPathRepos('Medias', repoName));
+							await ftp.upload(path[0]);
+							try {
+								await ftp.delete(media.old);
+							} catch (err) {
+								logger.warn(`File ${media.old} could not be deleted on FTP`, { service });
+							}
+						} else {
+							await ftp.rename(basename(media.old), basename(media.new));
+						}
+					}
+				}
+				await ftp.disconnect();
+			}
 			emitWS('pushComplete', repoName);
 		} catch (err) {
 			throw err;
